+++ test/java/JavaBlockMicrogrammarTest.ts
@@ -16,7 +16,7 @@ describe("JavaBlock microgrammars", () => {
 
 });
 
-const JAVA_IDENTIFIER = /^[A-Za-z][a-zA-Z0-9]+/;
+export const JAVA_IDENTIFIER = /^[A-Za-z][a-zA-Z0-9]+/;
 
 const METHOD_GRAMMAR = Microgrammar.fromDefinitions<ChangeControledMethod>({
     _visibilityModifier: "public", // TODO make optional